Log actions to database using the XMLRPC link.
The begin/end logging calls are inserted around the action exec. For
this wrapping to be in the common codepath, define an action method
in the svc class. rcService and svcmgr now use that new method instead
of direct lookup of the action in the service attributes.

Misc fixes alongside:
1/ finish half baked transition of svcmgr to svcBuilder's get_services
2/ reindent svcmgr python-style