Add dds sync resource.
dds is a tool to extract a Linux LVM2 snapshot (dedicated)
copy-on-write table and the associated changed dataset to a
stream or a file.

OpenSVC drives dds runs as a normal per-lv sync resource:

1/ a snap1 is constantly attached to lv
2/ upon sync
2.0/ check if snap1 uuid matches the last sync uuid on remote
2.1/ a new snap2 is attached to lv
2.2/ the cow table is read from snap1
2.3/ the dataset is extracted from snap2
2.4/ snap1 is destroyed
2.5/ snap2 is renamed snap1
2.6/ new snap1 uuid is stored on the source and destination nodes.

Supported actions:

1/ syncfullsync
2/ syncupdate

Snap size defaults to 10% of its origin lv, and can be ajusted
through the 'snap_size' env file parameter.