fix warning 'can not stat /dev/dm-2' on Linux mount resources
This unwanted warning could appear when listing the devices of a
mount resource (ie for every action) when there are snapshots of
a device held by the mount.

The realdev is a devmap but it's intentionally prevented from having
a device node in /dev to discourage bypassing the snapshot code
when accessing the origin.