Add vdisk resource
vdisks are used to remap virtual machines disks.

[vdisk#1]
path@node1 = /dev/svc1/rootdisk
path@node2 = /dev/vg0/svc1_rootdisk

with vdisks one can use different files or devpaths as vm disk
depending on the node. The typical situation is a dedicated vg on
production nodes and a shared vg on the drp node.

to accomodate this particular scenario, this patch also introduce
'disable_on = node1 node2' configuration parameter. This can be used
to disable a resource (the vg in this scenario) on the drp node.