Remove stack printing from 'except ex.excError'
if the exception is converted to ex.excError, it's a known error
path. Stack printing is reserved for unexpected exceptions.

For example this stack printing was triggered when a zfs receive
failed. In this case a proper error is printed from the zfs command
on stderr : no pertinent information is displayed in the stack.