Print status and log to collector when container ips are not resolved
The behaviour up to now was to abort all actions, which prevented
service config/status push to collector. The problem couldn't be
reported from the collector interfaces, and worse the service finally
disapeared from the collector when considered obsolete.

Also rework the way the logger is initialized so that we never end
up with multiple handlers for the same sink.

And teach the lxc container resource to use its parent _status()
method instead of its inferior private one.

Fix svc build exception trapping, where we used a wrong exception
reference