Add virtualbox support
OpenSVC now handles services of the vbox type (mode=vbox).
Tested on Darwin, but should work on other systems where virtualbox
is supported.

Keep in mind
a/ the vm are run by root (and vm started by a different user are not
seen by root)
b/ NATed VM are not pingable from the host, so avoid this mode for
serious usage