Add com.opensvc.sys.group and opensvc.sys.group.membership compliance modules
Those are split because users and group must exist to manipulate
group membership information.

Support user and group creation if the 'db' property is not set or
not set to 'nis', 'ldap', 'ad', or whatever remote account database.

Group variables are prefixed with 'group_' in the collector's
compliance rules.

Example group value:
{"opensvc": {"gid": 1003, "members": ["cvaroqui"]}, "test": {"gid": 1010, "members": ["test1", "test2"]}}