Add flex clusters configuration parameters
flex are n-way active clusters. The collector monitors how many
nodes are active with the knowledge of a min and max nodes, and
monitors the average cpu usage across those active nodes with
the knowledge of low and high thresholds.

The collector alerts upon out-of-range min/max nodes and low/high
cpu usage. Autoflex clusters allow the collector to take action to
keep these metrics in-bound, by starting/stopping service instances.

The implementation target a n-nodes farm like this :

---------------------------------------------------------------
| svc  |                     nodes                            |
---------------------------------------------------------------
|      | svc1      | svc2      | svc3      | svc1, svc2, svc3 |
|      | dedicated | dedicated | dedicated | shared           |
|      | nodes     | nodes     | nodes     | nodes            |
---------------------------------------------------------------
| svc1 | XXXXXXXXX |           |           | X                |
| svc2 |           | XXXXXXXXX |           | XXXXXXX          |
| svc3 |           |           | XXXXXXXXX | XX               |
---------------------------------------------------------------

The best service mode to do that is zone or linux containers as
they present the lowest performance degradation and keep the
concurrent applications setups isolated.