Variable frequency stats pushing
Variable frequency is needed by flex farms, whose alarming is
partly based on stats metrics ... those better be up-to-date
for the collector to report pertinent alerts, and to take
appropriate decisions in the case of autoflex clusters.

Reference implementation on Linux. All others need conversion, and
are actually broken for the next few days.

Introduce the 'nodemgr --stats-interval' parameter to control the
stats backlog extent we want to push. Next step will be a node
configuration file to put the interval into.

Dev note: the stats parsers are now inheriting a common class to
factorize some low-level function that were previously duplicated.
Also all stats are pushed in one xmlrpc call, and the data is
transmitted as a cPickle string to avoid xml markup verbosity.

Better, more maintainable and faster.