'autostart_node' evolution
Now autostart node can take a whitespace separated list of nodes,
subset of the 'nodes' list.

Failover cluster are still expected to have only 1 autostart_node,
but flex can hapilly specify a list of node to start the service
on.

Also, the flex considered the autostart_node as the replication
master. Not anymore. Now the 'flex_primary' param points this
replication master.