com.opensvc.sys.files compliance module enhancement
extend the dictionnary possibilities:

'ref':
======
points to a http://.. url to fetch the reference file from.
'ref' is exclusive with 'fmt'. Referenced files can embed
%%VAR%% and %%ENV:VAR%% wildcards, like 'fmt'.

'mode':
=======
Octal mode representation of the file to check, as a string.

'uid':
======
user name of the file ownership.

'gid':
======
group name of the file ownership.