Add node authentication mecanism
Add 'nodemgr register' command to ask the collector for a uuid.
The uuid is stored in node.conf and the tuple (nodename, uuid) is
provided upon every xmlrpc call.

The collector verifies this tuple match its view. If not the xmlrpc
returns immediately (before changing or fetching any data).

The avoid spoofing attacks, you are advised to move you clients to
the https feeder, using in node.conf:

[node]
dbopensvc = https://feedopensvc.opensvc.com/feed/default/call/xmlrpc
dbcompliance = https://collector.opensvc.com/init/compliance/call/xmlrpc

The nodeware detects if the collector provides the node auth facility,
so that this version is compatable with pre-node-auth collectors.
Users of collector.opensvc.com are encouraged to upgrade their node as
soon as possible so that collector.opensvc.com can be upgraded for
node auth.