If dbopensvc or dbcompliance is set, update rcEnv url fields
Those field are used to test the collector pings for one. So
whitout this patch, the nodeware check the internal default host
instead of the one specified in the node.conf url.

With this patch you don't have to set 'dbopensvc' in your dns or
host file anymore if the url are set in node.conf.