Add OVM heartbeat resource
An 'ovm' service with a 'ovm' heartbeat describes a Oracle VM with
failover enabled at the manager level.

In this case the heartbeat resource takes care of disabling the
failover on the manager upon 'stop' and re-enabling upon 'start'.

The ovm heartbeat status monitors OVS process presence and current
status of the failover attribute on the manager.

This patch also moves to cluster safety net to svc::action() to
prevent more actions.