Loopback mounted fs enhancement on Linux
If multiple fs with the same file backend are mounted, the system
will allocate different loopback devices for each mount. This is
not optimal and losetup will report only the first loop/file
binding. Which makes things difficult to detect if the mount is
up.

This patch detects upon mount start if the file backend already
has a loopdev bound. If so, use the loopdev and drop the 'loop'
mount option.