Be more specific in OpenHA daemons presence testing
The OpenHA resource driver only tested the presence of
nmond, heartc and heartd daemons. This is wrong as heartc
and heartd are not launched if the user defines a single
disk-based heartbeat.

This patch make the driver deduce the exact names of the OpenHA
daemons to check from the monitor configuration file.