Implement 'svcmgr disable' and 'svcmgr enable' commands
The disable/enable commands required a --rid argument. This
patch makes this argument optional. When no rid is specified
the disabling/enabling applies to the whole service.

Service disabling is required to restart the OpenHA heartbeat
without touching the services and preserving the heartbeat
service status.

Also fix the disable/enable/create/delete commands on the
service symlink, which unduly stacked.