Fix nodemgr help message
The service compliance patch broke the nodemgr help message: it
would not display the compliance commands anymore.

This was because we tested the command to be a method of the node
object, but the compliance methods moved to another object for
factorization with the service object.

This patch adapts the test to search for compliance methods in
the compliance object.