Add "reboot" and "shutdown" nodemgr commands
Linux and HP-UX implementations.

Those are for the collector to be able to mass-shutdown (typical
emergency action in reaction to a partial power loss in a
datacenter).

For that implementation, introduce lib/node{sysname}.py modules,
containing OS-specific code. Also introduce ximport helper, which
looks up "base+sysname" modules, and falls back to "base".