Add a raw vg resource
Defined by

[vg#0]
type = raw
devs = /dev/sdb /dev/mapper/60000000000000000001
scsireserv = true

This resource is meant for opensvc to handle scsi reservations
for a bunch of devices used as-is by the service. For example
raw devices used by a database.

Also support the 'type' parameter as a preferred alias for 'vgtype'
to align with the other resources 'type' parameter usage.