Add sysctl compliance object
Take json serialized data like:

[
 {"key": "net.unix.max_dgram_qlen", "value": [">=", 9]},
 {"key": "kernel.ctrl-alt-del", "value": ["=", 1]},
 {"key": "kernel.printk", "value": [[], [] , [], [">=", 12]]}
]

In the case of kernel.printk, typical value is a b c d. Leaving
the first 3 subvalues as [] tells the object to ignore a b c and
just check d is >= 12.

If d is not >= 12, a b c values are grabed from the live
kernel.printk value and d is replaced by 12.