Randomize cron jobs to level collector's load
This randomization does not apply to sync jobs, which
need to be triggered precisely to guarantee data
coherence.

            proba
              ^
        100%  |
         75%  |X
         50%  |XX
         25%  |XXX
          0%  ----|----|-> elapsed
             0%  50%  100%

The idea is to skip 75% of actions in period's first run,
skip none after half the interval is consumed, and decrease
skip probability linearly in-between.

This algo is meant to level collector's load which peaks
when all daily cron trigger at the same minute.