Fix the pushed stats interval computation
Now that the push_interval is set to 121 minutes, the algo
computing the range of statistics to push to the collector is
wrong : it pushed 242 minutes of stats.

Fix it to report from the start of the previous day allowed
push_period.

Also rename the --stats-start to --begin and --stats-end to --end
to align to other command range specification options.