service 'boot' action fix
If a primary node boots while the service is up on a secondary
node, the service 'boot' action would try a service 'start' which
fails on conflicting resource.

All is well, except if the service has standby resources defined,
they would not be start.

This patch implement startstandby() on failed start() in the
boot action.

example:

[CMI]n4:/opt/opensvc# drptestdev01.consulting-cmi.com boot
10:14:59 INFO  TESTDEV01.IP#1  checking 188.165.43.172 availability
10:14:59 ERROR TESTDEV01.IP#1  188.165.43.172 is already up on another host
10:14:59 INFO  TESTDEV01       start failed. try to start standby
10:15:00 INFO  TESTDEV01.FS#1  e2fsck -p /dev/mapper/vg0-drptestdev01
10:15:00 INFO  TESTDEV01.FS#1  output:
/dev/mapper/vg0-drptestdev01: clean, 244977/1966080 files, 4711016/7864320 blocks

10:15:00 INFO  TESTDEV01.FS#1  mount -t ext4 -o defaults /dev/mapper/vg0-drptestdev01 /drptestdev01