Resource action triggers enhancement
Allow (pre|post)_<action>[@nodes|@drpnodes|@<nodename] syntax for
every service resource.

And finally document this useful feature.

'pre'-triggers are a convenient way for a service to wait for another
service. For example, a service using a nfs share can wait for the
service serving this nfs share. With such a setup, rebooting both
client and server cluster would succeed.