Merge user dictionnaries gathered from different rulesets
For example, a ruleset may define uid, gid, shell and
another ruleset define the passwd for the same user. In
this case the users comp object overwrote the first
dict with the second.

This patch replace this overwrite logic by a merge of
the distinct keys. In case of key conflict, the first
evaluated key value is kept.