Fix HP-UX scsimgr cache
rcDiskInfo implements a cache of scsimgr information, but the
index of this hash was not appropiate.

This patch indexes the hash with all possible device alias to
minimize the cache miss.

Now that the cache hit is 100%, also convert device size to
megabytes, as was already done in the cache miss code path.

With this patch, the "nodemgr pushdisks" is down from 2min to
10sec on HP-UX servers with about a hundred paths.