Add Brocade switch configuration push to collector
'nodemgr pushbrocade' pushes a switchshow command
output to the collector. The parser will be on the
collector, as is done for storage array parsers.

The pushbrocade nodemgr command depends on the node
being ssh-trusted on the switchs it will collect.

The switchs and credential are pointed in auth.conf
with sections like

[1.2.3.4]
type = brocade
username = admin
key = /root/.ssh/id_rsa