Add 'nodemgr updatepkg' command
This new command updates the opensvc package on the node.

It depends on a node.repo node.conf parameter, pointing to
a ftp or http repository hosting opensvc packages.
The target package must be defined as 'current'.

The repo file tree must comply to the following organization:

ROOT
+- packages
 +- pkg
  +- opensvc-1.4-1.pkg
  +- opensvc-1.4-2.pkg
  +- current -> opensvc-1.4-2.pkg
 +- rpm
 +- deb
 +- depot
 +- tbz

The package fetching code is cross os, and the update methods are
os specific. This first patch implements the SunOS update method.