Add rcIfconfig hwaddr parameter on HP-UX
On HP-UX the hardware address is not displayed in ifconfig. Add
an optional hwaddr toggle to rcIfconfig object that triggers an
additional lanscan command parser.

The ifconfig parser looks up the interface name in a hash
populated by the lanscan parser.

This patch is needed for 'nodemgr pushasset' to report the lan
configuration to the collector on HP-UX.