Allow the rsync resource to login to remote nodes with specified user
The rsync resource use root as the default remote user to login
on the remote cluster nodes. Some sites don't allow root logins
at all.

Setting 'ruser' in node.conf tells the rsync resource to use
an alternate remote user instead of root.

The remote user must have the privileges to run as root the
following commands on the remote node:
 - nodemgr
 - svcmgr
 - rsync

The default ruser is root for all nodes. ruser accepts a list of
user[@node] ... If @node is ommited, user is considered the new
default user

Examples:

;ruser = opensvc
;ruser = root opensvc@node1
;ruser = usr1@node1 usr2@node2 usr3@node3