Add new tags to Solaris zone ip resources
* exclusive

this tag tells opensvc to handle an ip resource for exclusive-type
zones. ifconfig are fetched in the zone instead of in the global
to determine if the ip is up on ipdev.

* noalias

tells opensvc not to stack an ip alias but to plumb the ip on
the base ethernet interface. in this case, the netmask parameter
must be set for the resource as opensvc can not deduce the
netmask from the interface primary ip.

* actions

by default, opensvc won't take actions (stop, start) on ip resources
tagged as exclusive, as the guest os can be configured to bring
this ip by himself at boot. this tag instructs opensvc to take
actions, overriding the default behaviour.

These new features depend on some core ip resource changes:

- rcIfconfigSunOS.ifconfig() can now take an ifconfig output as
argument. In this case, the ifconfig in the global is not
captured. The resIpZone implement passes to ifconfig() the
output of a ifconfig grabbed from the zone for ip resources
in 'exclusive' mode.

- normal zone ip resources (no tags) are started after zone
ready. 'exclusive actions' resources are started after zone
boot.

- as the network is not fully started when the zone boot command
returns, a wait loop is implemented (maxing at 30 seconds) to
wait for the network to be brought up before starting ip
resources flagged 'exclusive actions'

- disable arp announce for 'exclusive' zone ip resource