Restore the startip action behaviour on zones
A previous patch made the 'actions' tag mandatory to execute any
action on a zone ip. This changed the behaviour for non-exclusive
zones.

This patch restores the previous behaviour for non-exclusive zones,
ie do actions even if the 'actions' flag is not set.

The default for exclusive zones is still to not trigger ip actions
if the 'actions' flag is not set.