Add cluster service parameter and clusters node parameter
A node can be a member of multiple overlapping clusters. In this
case, the node.clusters parameter would be set to a whitespace
separated list of cluster names, and each service would have to
set its own single cluster name using the DEFAULT.cluster
parameter.

If the node is member of a single cluster, the service cluster
name is automatically set to the node cluster name.

These parameters are optional, and are used only to report the
cluster name as the virtual array name when registering VM
virtual disks.