Fix stack in Dcs snap ressource
This stack happened when a snapshot defined in the
ressource configuration does not exist on the dcs head.
The get-dcssnapshot does not return an error in this
case, so we use the information dictionniary 0-length
as a hint that the snapshot is not present and return
the expected None value from get_snap() instead of the
empty dictionnary.