Support partition to disk device translation on Linux
In disklist() function on Linux, we try to find the disks
each ressource relies on. This is used in scsi persistent
reservation to identify which disks need to be reserved.

The vg ressource disklist() stopped the resolution at
the pv level when the pv is not a device map. This
caused disklist to return partitions as disks. As
partitions are not reservable, this caused service
startup errors.