Add connection id management to Datacore commands sessions
The Datacore manager sessions clash if the user does not
set a connection id, because of its simplistic choice of
implicit connection id.

Generate a connection from opensvc as a md5 hash of the
service name and a uuid.

For out-of-service operations, like disk provisioning or
datacore configuration extractors, the connection id is
only uuid-based.