Change flex_max_nodes defaults to the number of autostart nodes
In a flex service, all nodes we want to start an instance of the
service on must be listed in DEFAULT.autostart_node

Use that as a default for the flex_max_nodes threshold, so that
the collector can raise an alarm if a n+1th instance is started.

The previous default value was the number of nodes (not including
the DRP nodes), but some flex are configured to have their DRP
instances started.