Fix env reset for "svcmgr compliance ..." loop
Popen doc says "If env is not None, it must be a mapping
that defines the environment variables for the new process;
these are used instead of inheriting the current process
environment, which is the default behavior."

But actually Popen seems to cache the env upon first call,
which defeats the env reset logic implemented in OpenSVC
between services compliance runs.

Explicitely pass env=os.environ to workaround the Popen
behaviour.