Misc fixes for regression due to vm-as-service
- add rcp() implementation to lxc container, based on cp

- add error="ignore|raise" kwarg to Svc::encap_cmd(). Defaults to
"raise". If set to "continue", the encap_cmd loop on containers
move on to the next container. If verbose is also set, a warning
is emitted.

- set error="continue" in Svc::stop()

- trap encap_cmd errors when fetching an encap service json status

- add 'boot' to the list of actions not needing to specify --master
or --slave or both

- fix startstandby for encap services: don't try encap_rmd if
container is down and always_on is False

- fix the _master_action/_slave_action decorators to validate
the action called by the user (and not the decorated function name)
is in the action whitelist.