Implement a generalized service start abort mecanism
Each resource can declare a abort_start() method that returns
True to abort the service start.

This is a generalization of the check_ip_ping_raise abort
mecanism.

New user of this generalized mecanism is the container resource
which aborts the service start if the container is started on
another node of the cluster.

The ip resource is ported to this mecanism.