Don't execute 'ctsm info' on boot on HP-UX
On HP-UX, the mem banks reporting depends on cstm. To
report an uptodate value, cstm is ordered to refresh its
maps. But this operation happens to hang for 15min when
executed on boot, and causes ServiceGuard to TOC.

Teach pushasset not to refresh those maps on boot. This
verification algo computes the difference between now
and 'who -b' date, and shortcuts the refresh if this
delta is less than 1h.