Fix hang on xmlrpc server methods fetching
If the xmlrpc server methods cache is unitialized or corrupted,
the agent try to fetch the lists from the feed xmlrpc service
and from the compliance xmlrpc service. These operation hang the
run for a very long time when the collector is not available.

This patch makes sure we never try to fetch the feed xmlrpc
method list when node.conf::node.dbopensvc = None, and never try
to fetch the compliance xmlrpc method list when
node.conf::node.dbcompliance = None

Beware, dbopensvc and dbcompliance have default values different
from None.