Add Veritas Cluster OpenSVC service wrapper (Linux)
[DEFAULT]
mode = veritas
pkg_name = DataService
service_type = DEV

This simple env file tell opensvc to wrap a service over a
Veritas package.

- The nodes list and heartbeat resource are fetched from hagrp
- The resource list is fetched from hagrp -resource
- The monitored flag is set according to hares Critical prop
- The resource status is fetched from hares State prop

'stop' and 'start' actions are not supported.
'print *' work as expected, as is the disk inventory and
compliance commands.

The ip and volume resources are not yet implemented.
The fs and vg resources are.

dataservice.linvoice.com
overall                   warn
|- avail                  up
|  |- vg#vcs0        M... up       datadg
|  |- fs#vcs0        M... up       /dev/vx/dsk/datadg/web@/web
|  '- app            .... n/a      app
|                                  # no checkup scripts
|- sync                   down
|  '- sync#i0        .... down     rsync svc config to drpnodes, nodes
|                                  # rh63tst03.linvoice.com need update
'- hb                     up
   '- hb#sg0         .... up       DataService