lvremove logging tweaks on Linux LVM2
lvremove may log those messages on stderr, which get reported to
the collector as error:

- Node /dev/mapper/vg0-osvc_sync_cmiprdweb01 was not removed by udev.
Falling back to direct node removal.

- The link /dev/vg0/osvc_sync_cmiprdweb01 should have been removed
by udev but it is still present. Falling back to direct link removal.

Those messages are informative, and should not be logged as errors.

Also suppress altogether the message about non-backed up metadata
update. We know what we do here. The snapshots we create are
transient.