Add node.conf 'compliance.auto_update' parameter
Boolean. Default False.

If set to True, and if the execution context indicates a scheduled run,
execute 'updatecomp' upon 'compliance check'.

This toggle helps keep the compliance modules in sync with the reference
repository. Beware of the security impact of this setting: you must be
careful your module repository is kept secure from malevolents.