Add 'nodemgr dequeue action' command
The collector now can either push command to the agents
(default on Unix) or queue the command and ask the
agent to poll the queue (default on Windows).

This new 'nodemgr dequeue action' command implements
the "pull" mode. The agent gets the queued commands from
the collector (which state are transitionned from waiting
to pending), executes the command either with nodemgr or
svcmgr, and finally submit the resulting returncode,
stdout and stderr tuple to the collector. The queue entries
are updated by the collector with this information, and
their status is transitionned from "pending" to
"terminated".