Add "zpool import -d" support
Zpool imports on servers with many disks can be very long.
The "zpool import" command takes a '-d' flag to point a directory
with symlinks to the pool block devices, so that it is not needed
to scan all devices for their pool ownership.

This path implements a zpool import try with -d
/opt/opensvc/var/<svcname>/dev/dsk/, and a fallback to the simple
import command in case this cache directory does not exists or
its content is not up-to-date.

Also adapt the rcDevTree and zpool checker for this case.