Don't run check_ping twice on service start
Leave a flag to remember when we do it the first time, and check
this flag on the second time before deciding we need to do check_ping.

Also shorten the timeout to 1 second on check_ping after stopping
the ip. The stopped ip being local, no need to wait too long.