Add btrfs device error checker
This checker pushes to the collector the stats exported through
'btrfs dev stats <mount>'.

This checker is actually the most practical way of alerting that
a btrfs might need a scrub.

After redundancy is restored, the 'btrfs dev stats -z <mount>'
can be used to reset the stats so that the alert can be cleared
in the collector.