Lxc container driver enhancements
1/
Allow users to specify the container config file location through
the 'cf' resource keyword

2/
Reorganize the code so that the container config file can be hosted
by a fs resource driven by the service.

3/
A lxc container resource contributes the container fstab config file
to the sync#i0 resource, if the container config file defines the
'lxc.mount' param, and the pointed file actually exists.

4/
Containers accessed through lxc-attach don't need to wait for the
pingable status to go on with the service startup. To that end,
set a dummy getaddr() resource function, in addition to the already
implemented dummy ping() function.