Skip encap service start on containers in autostart_node
If a container hostname is in autostart_node, the encap service
will start from '/etc/init.d/opensvc start', so no need to
start it one more time from the master.

This patch adds the impersonate parameter to scoping parser
function, so that the master can determine the autostart_node
param value from the guests point of view.