More resApp fixups
1/
Check if script is executable later, on action. This check was
done on resource object attachment to the service object. At this
time, the script is generally not available yet in a startup
sequence.

2/
Fix a regression in fs mount/umount ordering introduced by the
previous patch.

3/
Fix a stack when the startup script is not available on stop action