SysV apps fixes
- don't stack on validate_script_exec() if self.script is None

validate_script_path() is allowed to set self.script to None.
validate_script_exec() is run in sequence, and should cope with
self.script as None.

- rescan app resource from sysv definitions when the app resourceset
action() is called

At service build time, the sysv app launchers may not be visible
if they are hosted in a fs resource handled by the service. In this
case, provision an empty app resourceset with the rescan plugged
early in the action() function

- fix sysv vs cf resource matching when the script param value is a
basename

We always evaluated the sysv script as already defined in the config
file in this particular case due to some a==a bogus test.