Add new access methods for 3par arrays
The array section in auth.conf now requires the 'method' parameter.
Possible values are :

* ssh

This is the default mode prior to this patch: the agent submit commands
over ssh directly to the array. This method requires the 'manager',
'username' and 'key' parameters.

* cli

This method submit commands through the HP 3par cli toolkit. The cli
command must be in the PATH. This method requires the 'pwf' parameter.

* proxy

This method submits commands to a OpenSVC owned rest proxy. The proxy
has the HP 3par cli toolkit installed and relays commands if they
match a whitelist attached to the (node uuid, svc name, array name)
tuple.
This method requires the 'manager' parameter.
This method is tailored for DMZ OpenSVC agents with the arrays in the
internal networks.

Please contact OpenSVC to request this rest server.