Subsetting fix
Now that any resource set can be configured to spawn threads for
each resource through subsetting, we face 2 problems:

1/
Uncatched exceptions in threads are printed on screen.
Add a resource action wrapper to catch all exceptions in the thread
and translate to indicative exitcodes.

2/
The resource property changes are lost with the thread.
Use exitcode(2) to tell the parent process to set can_rollback
resource property.