Support ipmp path switching in Solaris ip resource
ipmp fails over ip aliases between member ifaces of a group when
the active member fails.

Teach the ip resource to look for the addr on all group members on
stop, and to look for a valid member on start.

The outputs are as usual when the primary member is up, and the
following messages are logged when the primary member is down:

root@sol1:/opt/opensvc$ solp26.opensvc.com startip
23:27:38 INFO    SOLP26.OPENSVC.COM.IP#0    switch rtls0 to valid alternate path rtls1
23:27:38 INFO    SOLP26.OPENSVC.COM.IP#0    checking 37.59.71.26 availability
23:27:40 INFO    SOLP26.OPENSVC.COM.IP#0    /usr/sbin/ifconfig rtls1:2 plumb 37.59.71.26 netmask + broadcast + up
23:27:40 INFO    SOLP26.OPENSVC.COM.IP#0    output:
Setting netmask of rtls1:2 to 255.255.255.224

root@sol1:/opt/opensvc$ solp26.opensvc.com stopip
23:27:50 INFO    SOLP26.OPENSVC.COM.IP#0    found 37.59.71.26 plumbed on rtls1:2, in the same ipmp group than rtls0
23:27:50 INFO    SOLP26.OPENSVC.COM.IP#0    /usr/sbin/ifconfig rtls1:2 unplumb
23:27:50 INFO    SOLP26.OPENSVC.COM.IP#0    checking 37.59.71.26 availability