Support globing in raw resources
When the raw resource is used for inventoring devices, it is
useful to set devs as a list of globing patterns to catch new
devices added to the service with no service config maintenance.

Note however that using this new feature voids the missing device
detection in resource status evaluation.