Remove icmp check of collector
To simplify the setup of the opensvc infrastructure in firewalled
environments, remove the agent to collector icmp check.

Also remove the /opt/opensvc/var/collector cache maintenance, as
we now use the xmlrpc methods fetch function as the collector
availability marker.

Finally, setup INFO as the default logging level instead of debug,
and set a proper loglevel to informative messages.