Try harder to get docker container id on start
If dockerd is not already running, we failed to get the
container id on resource attachment to the service, and never
retried after starting the daemon.

This caused the agent to try to run the image instead of starting
the existing container. In this case the docker daemon complains
about a container name conflict.