Use resource status cache files to speed up svcmon and print status
Also store the resource status string in the cache file, so that
this information is also available for display through print
status

"svcmon --updatedb" still ignores the caches so that status cache
gets updated on schedule.

Service actions also purge the caches so that decisions are taken
on up-to-date status information.

Service actions also refresh the cache after the action.

Add svcmon and print status colorization.