Don't stop app resources in down-state on service stop and shutdown
Which is the common resource behaviour.

Note that an app resource with no checker defined returns n/a as a
state. In this case we have no way to determine if the app stop
should be done or not. Thus, we have no choice but do the app stop,
be it usefull or not.