Implement a ping check for containers
If the container name is resolvable, prepend a ping check to
abort_start checks.

The container up state detection by asking the peer nodes through
ssh is not reliable enough to prevent double starts when a
parallel start of tens of services saturate the remote sshd.

The ping test, when possible, is faster and more reliable.