Implement start/stop/switch openha hb actions
This patch disables the --cluster safety net when a heartbeat
implements the start/stop/switch submitted action.

The heartbeat actions run first, and all make use of the new
excEndAction() exception to force the agent not to run the other
resource actions, as they are run within the context of the
--cluster actions triggered by the heartbeat daemon.

This patch also introduce the --remote flag to offer a generic
mean to avoid ping-ponging remote actions. The freeze/thaw
action now use this flag instead of the --force improper usage.

The switch action does not need the --to flag, as openha
cluster are currently limited to 2 nodes topologies.