Compliance module autoremediation support
A module can be set "autofix" through the collector interfaces,
in the context of a moduleset.

The scheduled action is no longer "compliance check", but
"compliance auto", which decides module per module if a check or
a fix must be done.

If a module appears in different attached modulesets, being
autofix in one is sufficent to trigger the fix.

This patch also makes all compliance run preparation using just
one rpc. "show moduleset" is also down to one rpc.

Warning, you need an up-to-date collector to run compliance action
with this patch.