Add the 'nodemgr sysreport' command
This command collects files and commands outputs on the node and
send them to the collector.

The collector commits them in a git node-dedicated file tree, and
offer interfaces to highlight and drill down changes, to compare
nodes configuration, ...

The agent as a minimalist set of files to collect. This set can
be extended through /opt/opensvc/etc/sysreport.conf.d/
configlets.

The configlet synthax is described in usr/share/doc/template.sysreport.conf

The sysreport node command is scheduled once a day. The [sysrepor]
node.conf section can be used to alter the default schedule.