Compliance enhancements
1/ Support moduleset encapsulations.

show moduleset now displays arbitrarily deep moduleset
encapsulations

2/ Each module has now a custom env setup.

The env is composed of :
- explicit rulesets attached to the node or service
- contextual rulesets
- explicit rulesets inherited through the attached moduleset
providing the module

The first case will tend to be exceptional.
Even the second case will be converted to the third case in most
situations.

As a consequence, the consumed env space is reduced to the
variables and their values necessary to the module role. This
solve the 130k env limit size on some systems and bring a much
better module-rules dependencies readbility.