Sysreport enhancements
- deprecate GLOB, FILE, DIR in favor of a single INC keyword in
sysreport configlets. INC means include and its value is a glob
pattern. The pattern is expanded and each path is auto detect as
a file or a dir. If its a dir, the whole content is added to the
sysreport

- add the EXC keyword. The same expansion logic as INC is applied.
The EXCs resulting file path set is substracted from the INCs set.

- GLOB, FILE, DIR are internally converted to INC for backward
compatibility