Add a sudoers compliance object
This object behaves and accepts the same variables as the files
compliance object.

It only implements an additional "visudo -c" check of the
delivered sudo rules before accepting to install or modify the
target file.

It is an interesting check considering a single broken rule breaks
the whole sudo configuration of a system.