Force hb status to up after a successful hb-triggered service startup
The hb resource of a service would be 'warn' until the next resource
status evaluation after a hb-triggered startup, because at the end
of the start action, the hb still has the service in STARTING state.

If the opensvc service startup is successful, force a 'up' state
to the hb resource.