Do the app checks if there are no reference resources
The agent would skip doing app resource status evaluation if
the service's reference resources are not up. A reference
resource being any resource except hb, sync, app and scsi
reservation.

In case there are no reference resources, this heuristic is
not appropriate, so this patch adds a reference resource count
and enables app resources check if this count is 0.

So we can now have minimalist services like:

overall                   up
|- avail                  up
|  '- app#1          .... up         redis_init_script_6379
|- sync                   up
|  '- sync#i0        .... up         rsync svc config to drpnodes, nodes
'- hb                     n/a